+++ app/manage/forms.py
@@ -6,7 +6,7 @@ from flask_wtf import FlaskForm
 from wtforms import StringField, BooleanField, SelectField, SelectMultipleField, SubmitField
 from wtforms.validators import InputRequired, Optional, MacAddress
 from ..models import Room, DeviceType
-from ..models import Lesson
+from ..models import LessonType
 
 
 class NewDeviceForm(FlaskForm):
@@ -16,7 +16,7 @@ class NewDeviceForm(FlaskForm):
     device_type = SelectField('设备类型', coerce=str, validators=[InputRequired()])
     room = SelectField('所属场地', coerce=str, validators=[InputRequired()])
     mac_address = StringField('MAC地址', validators=[Optional(), MacAddress()])
-    lessons = SelectMultipleField('授权课程', coerce=str)
+    lesson_types = SelectMultipleField('授权课程类型', coerce=str)
     development_machine = BooleanField('开发用机器')
     submit = SubmitField('提交')
 
@@ -27,7 +27,7 @@ class NewDeviceForm(FlaskForm):
         else:
             self.device_type.choices = [('', '选择设备类型')] + [(str(device_type.id), device_type.name) for device_type in DeviceType.query.filter(DeviceType.name != 'Server').order_by(DeviceType.id.asc()).all()]
         self.room.choices = [('', '选择所属场地')] + [('0', '无')] + [(str(room.id), room.name) for room in Room.query.order_by(Room.id.asc()).all()]
-        self.lessons.choices = [('', '选择授权课程')] + [(str(lesson.id), lesson.name) for lesson in Lesson.query.all()]
+        self.lesson_types.choices = [('', '选择授权课程类型')] + [(str(lesson_type.id), lesson_type.name) for lesson_type in LessonType.query.all()]
 
 
 class EditDeviceForm(FlaskForm):
@@ -37,7 +37,7 @@ class EditDeviceForm(FlaskForm):
     device_type = SelectField('设备类型', coerce=str, validators=[InputRequired()])
     room = SelectField('所属场地', coerce=str, validators=[InputRequired()])
     mac_address = StringField('MAC地址', validators=[Optional(), MacAddress()])
-    lessons = SelectMultipleField('授权课程', coerce=str)
+    lesson_types = SelectMultipleField('授权课程类型', coerce=str)
     development_machine = BooleanField('开发用机器')
     submit = SubmitField('提交')
 
@@ -48,4 +48,4 @@ class EditDeviceForm(FlaskForm):
         else:
             self.device_type.choices = [('', '选择设备类型')] + [(str(device_type.id), device_type.name) for device_type in DeviceType.query.filter(DeviceType.name != 'Server').order_by(DeviceType.id.asc()).all()]
         self.room.choices = [('', '选择所属场地')] + [('0', '无')] + [(str(room.id), room.name) for room in Room.query.order_by(Room.id.asc()).all()]
-        self.lessons.choices = [('', '选择授权课程')] + [(str(lesson.id), lesson.name) for lesson in Lesson.query.all()]
+        self.lesson_types.choices = [('', '选择授权课程类型')] + [(str(lesson_type.id), lesson_type.name) for lesson_type in LessonType.query.all()]