+++ app/manage/views.py
@@ -467,8 +467,8 @@ def device():
         )
         db.session.add(device)
         db.session.commit()
-        for lesson_id in form.lessons.data:
-            device.add_lesson(lesson=Lesson.query.get(int(lesson_id)))
+        for lesson_type_id in form.lesson_types.data:
+            device.add_lesson_type(lesson_type=LessonType.query.get(int(lesson_type_id)))
         db.session.commit()
         flash('已添加设备：{} [{}]'.format(device.alias, device.serial), category='success')
         add_user_log(user=current_user._get_current_object(), event='添加设备：{} [{}]'.format(device.alias, device.serial), category='manage')
@@ -643,9 +643,9 @@ def edit_device(id):
         device.mac_address = (None if form.mac_address.data == '' else form.mac_address.data)
         device.category = ('development' if form.development_machine.data and current_user.is_developer else 'production')
         db.session.add(device)
-        device.empty_lessons()
-        for lesson_id in form.lessons.data:
-            device.add_lesson(lesson=Lesson.query.get(int(lesson_id)))
+        device.remove_all_lesson_types()
+        for lesson_type_id in form.lesson_types.data:
+            device.add_lesson_type(lesson_type=LessonType.query.get(int(lesson_type_id)))
         db.session.commit()
         flash('已更新设备信息：{}'.format(device.alias2), category='success')
         add_user_log(user=current_user._get_current_object(), event='更新设备信息：{}'.format(device.alias2), category='manage')
@@ -656,7 +656,7 @@ def edit_device(id):
     form.device_type.data = str(device.type_id)
     form.room.data = ('0' if device.room_id == None else str(device.room_id))
     form.mac_address.data = device.mac_address
-    form.lessons.data = [str(item.lesson_id) for item in device.lesson_authorizations]
+    form.lesson_types.data = [str(item.lesson_type_id) for item in device.lesson_type_authorizations]
     form.development_machine.data = (device.category == 'development')
     return minify(render_template(
         'manage/edit_device.html',
@@ -692,8 +692,9 @@ def delete_device(id):
     device = Device.query.get_or_404(id)
     if device.category == 'development' and not current_user.is_developer:
         abort(403)
+    device.remove_all_lesson_types()
     db.session.delete(device)
-    flash('已恢删除设备：{}'.format(device.alias2), category='success')
+    flash('已删除设备：{}'.format(device.alias2), category='success')
     add_user_log(user=current_user._get_current_object(), event='删除设备：{}'.format(device.alias2), category='manage')
     db.session.commit()
     return redirect(request.args.get('next') or url_for('manage.device'))