+++ app/study/views.py
@@ -7,6 +7,7 @@ from flask import render_template, jsonify, redirect, request, url_for, abort, f
 from flask_login import login_required, current_user
 from . import study
 from .. import db
+from ..models import Device
 from ..models import LessonType, Lesson, Video
 from ..models import Punch
 from ..decorators import permission_required
@@ -85,7 +86,18 @@ def test_review():
 @study.route('/demo')
 def demo():
     '''study.demo()'''
+    mac_address = get_mac_address_from_ip(ip_address=request.headers.get('X-Forwarded-For', request.remote_addr))
+    if mac_address is None:
+        flash('无法获取设备信息', category='error')
+        return redirect(url_for('auth.login'))
+    device = Device.query.filter_by(mac_address=mac_address).first()
+    if device is None:
+        flash('设备未授权（MAC地址：{}）'.format(mac_address), category='error')
+        return redirect(url_for('auth.login'))
     lesson_type = '体验课程'
+    if not device.can_access_lesson_type(lesson_type_name=lesson_type):
+        flash('该设备无法访问受限资源', category='error')
+        return redirect(url_for('auth.login'))
     lessons = Lesson.query\
         .join(LessonType, LessonType.id == Lesson.type_id)\
         .filter(LessonType.name == lesson_type)\