+++ RELEASE-NOTES.md
@@ -9,7 +9,7 @@
 - **Model** - Add `restricted_permit` attribute to device data model #9
 - **Feature** - Only permitted devices can access restricted videos #10
 - **Feature** - Add VB demo lessons, Y-GRE demo lessons, and Y-GRE test reviews #11
-- **Model** - Add device lesson data model #13
+- **Model** - Add device lesson type data model #13
 - **Model** - Add `view_point` attribute to lesson type data model #19
 - **Model** - Add `login_required` attribute to lesson type data model #20